fix hydration error, other small fixes, cleanup
